// src/maocars-client/schemas/getAuctionsSort.ts
/* eslint-disable */
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * mao-cars-api
 */

export type GetAuctionsSort = (typeof GetAuctionsSort)[keyof typeof GetAuctionsSort];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetAuctionsSort = {
  lowest_mileage: 'lowest_mileage',
  newly_listed: 'newly_listed',
  no_reserve: 'no_reserve',
  charity: 'charity',
} as const;